requirements.txt
torch==1.11.0+cu113
torchmetrics==0.11.0
torchvision==0.12.0+cu113
timm==0.6.13
pillow==9.0.1
numpy==1.21.2
wandb==0.12.7